+++ examples/chain_client/distribution/query/9_CommunityPool.py
@@ -1,0 +1,15 @@
+import asyncio
+
+from pyinjective.async_client import AsyncClient
+from pyinjective.core.network import Network
+
+
+async def main() -> None:
+    network = Network.testnet()
+    client = AsyncClient(network)
+    community_pool = await client.fetch_community_pool()
+    print(community_pool)
+
+
+if __name__ == "__main__":
+    asyncio.get_event_loop().run_until_complete(main())