+++ pyinjective/async_client.py
@@ -9,10 +9,11 @@ from warnings import warn
 import grpc
 from google.protobuf import json_format
 
-from pyinjective import constant
 from pyinjective.client.chain.grpc.chain_grpc_auth_api import ChainGrpcAuthApi
 from pyinjective.client.chain.grpc.chain_grpc_authz_api import ChainGrpcAuthZApi
 from pyinjective.client.chain.grpc.chain_grpc_bank_api import ChainGrpcBankApi
+from pyinjective.client.chain.grpc.chain_grpc_distribution_api import ChainGrpcDistributionApi
+from pyinjective.client.chain.grpc.chain_grpc_exchange_api import ChainGrpcExchangeApi
 from pyinjective.client.chain.grpc.chain_grpc_token_factory_api import ChainGrpcTokenFactoryApi
 from pyinjective.client.chain.grpc.chain_grpc_wasm_api import ChainGrpcWasmApi
 from pyinjective.client.chain.grpc_stream.chain_grpc_chain_stream import ChainGrpcChainStream
@@ -174,6 +175,18 @@ class AsyncClient:
             ),
         )
         self.authz_api = ChainGrpcAuthZApi(
+            channel=self.chain_channel,
+            metadata_provider=lambda: self.network.chain_metadata(
+                metadata_query_provider=self._chain_cookie_metadata_requestor
+            ),
+        )
+        self.distribution_api = ChainGrpcDistributionApi(
+            channel=self.chain_channel,
+            metadata_provider=lambda: self.network.chain_metadata(
+                metadata_query_provider=self._chain_cookie_metadata_requestor
+            ),
+        )
+        self.chain_exchange_api = ChainGrpcExchangeApi(
             channel=self.chain_channel,
             metadata_provider=lambda: self.network.chain_metadata(
                 metadata_query_provider=self._chain_cookie_metadata_requestor
@@ -572,6 +585,455 @@ class AsyncClient:
         pagination: Optional[PaginationOption] = None,
     ) -> Dict[str, Any]:
         return await self.bank_api.fetch_send_enabled(denoms=denoms, pagination=pagination)
+
+    async def fetch_validator_distribution_info(self, validator_address: str) -> Dict[str, Any]:
+        return await self.distribution_api.fetch_validator_distribution_info(validator_address=validator_address)
+
+    async def fetch_validator_outstanding_rewards(self, validator_address: str) -> Dict[str, Any]:
+        return await self.distribution_api.fetch_validator_outstanding_rewards(validator_address=validator_address)
+
+    async def fetch_validator_commission(self, validator_address: str) -> Dict[str, Any]:
+        return await self.distribution_api.fetch_validator_commission(validator_address=validator_address)
+
+    async def fetch_validator_slashes(
+        self,
+        validator_address: str,
+        starting_height: Optional[int] = None,
+        ending_height: Optional[int] = None,
+        pagination: Optional[PaginationOption] = None,
+    ) -> Dict[str, Any]:
+        return await self.distribution_api.fetch_validator_slashes(
+            validator_address=validator_address,
+            starting_height=starting_height,
+            ending_height=ending_height,
+            pagination=pagination,
+        )
+
+    async def fetch_delegation_rewards(
+        self,
+        delegator_address: str,
+        validator_address: str,
+    ) -> Dict[str, Any]:
+        return await self.distribution_api.fetch_delegation_rewards(
+            delegator_address=delegator_address,
+            validator_address=validator_address,
+        )
+
+    async def fetch_delegation_total_rewards(
+        self,
+        delegator_address: str,
+    ) -> Dict[str, Any]:
+        return await self.distribution_api.fetch_delegation_total_rewards(
+            delegator_address=delegator_address,
+        )
+
+    async def fetch_delegator_validators(
+        self,
+        delegator_address: str,
+    ) -> Dict[str, Any]:
+        return await self.distribution_api.fetch_delegator_validators(
+            delegator_address=delegator_address,
+        )
+
+    async def fetch_delegator_withdraw_address(
+        self,
+        delegator_address: str,
+    ) -> Dict[str, Any]:
+        return await self.distribution_api.fetch_delegator_withdraw_address(
+            delegator_address=delegator_address,
+        )
+
+    async def fetch_community_pool(self) -> Dict[str, Any]:
+        return await self.distribution_api.fetch_community_pool()
+
+    # Exchange module
+
+    async def fetch_subaccount_deposits(
+        self,
+        subaccount_id: Optional[str] = None,
+        subaccount_trader: Optional[str] = None,
+        subaccount_nonce: Optional[int] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_subaccount_deposits(
+            subaccount_id=subaccount_id,
+            subaccount_trader=subaccount_trader,
+            subaccount_nonce=subaccount_nonce,
+        )
+
+    async def fetch_subaccount_deposit(
+        self,
+        subaccount_id: str,
+        denom: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_subaccount_deposit(
+            subaccount_id=subaccount_id,
+            denom=denom,
+        )
+
+    async def fetch_exchange_balances(self) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_exchange_balances()
+
+    async def fetch_aggregate_volume(self, account: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_aggregate_volume(account=account)
+
+    async def fetch_aggregate_volumes(
+        self,
+        accounts: Optional[List[str]] = None,
+        market_ids: Optional[List[str]] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_aggregate_volumes(
+            accounts=accounts,
+            market_ids=market_ids,
+        )
+
+    async def fetch_aggregate_market_volume(
+        self,
+        market_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_aggregate_market_volume(
+            market_id=market_id,
+        )
+
+    async def fetch_aggregate_market_volumes(
+        self,
+        market_ids: Optional[List[str]] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_aggregate_market_volumes(
+            market_ids=market_ids,
+        )
+
+    async def fetch_denom_decimal(self, denom: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_denom_decimal(denom=denom)
+
+    async def fetch_denom_decimals(self, denoms: Optional[List[str]] = None) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_denom_decimals(denoms=denoms)
+
+    async def fetch_chain_spot_markets(
+        self,
+        status: Optional[str] = None,
+        market_ids: Optional[List[str]] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_spot_markets(
+            status=status,
+            market_ids=market_ids,
+        )
+
+    async def fetch_chain_spot_market(
+        self,
+        market_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_spot_market(
+            market_id=market_id,
+        )
+
+    async def fetch_chain_full_spot_markets(
+        self,
+        status: Optional[str] = None,
+        market_ids: Optional[List[str]] = None,
+        with_mid_price_and_tob: Optional[bool] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_full_spot_markets(
+            status=status,
+            market_ids=market_ids,
+            with_mid_price_and_tob=with_mid_price_and_tob,
+        )
+
+    async def fetch_chain_full_spot_market(
+        self,
+        market_id: str,
+        with_mid_price_and_tob: Optional[bool] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_full_spot_market(
+            market_id=market_id,
+            with_mid_price_and_tob=with_mid_price_and_tob,
+        )
+
+    async def fetch_chain_spot_orderbook(
+        self,
+        market_id: str,
+        order_side: Optional[str] = None,
+        limit_cumulative_notional: Optional[str] = None,
+        limit_cumulative_quantity: Optional[str] = None,
+        pagination: Optional[PaginationOption] = None,
+    ) -> Dict[str, Any]:
+        # Order side could be "Side_Unspecified", "Buy", "Sell"
+        return await self.chain_exchange_api.fetch_spot_orderbook(
+            market_id=market_id,
+            order_side=order_side,
+            limit_cumulative_notional=limit_cumulative_notional,
+            limit_cumulative_quantity=limit_cumulative_quantity,
+            pagination=pagination,
+        )
+
+    async def fetch_chain_trader_spot_orders(
+        self,
+        market_id: str,
+        subaccount_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_trader_spot_orders(
+            market_id=market_id,
+            subaccount_id=subaccount_id,
+        )
+
+    async def fetch_chain_account_address_spot_orders(
+        self,
+        market_id: str,
+        account_address: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_account_address_spot_orders(
+            market_id=market_id,
+            account_address=account_address,
+        )
+
+    async def fetch_chain_spot_orders_by_hashes(
+        self,
+        market_id: str,
+        subaccount_id: str,
+        order_hashes: List[str],
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_spot_orders_by_hashes(
+            market_id=market_id,
+            subaccount_id=subaccount_id,
+            order_hashes=order_hashes,
+        )
+
+    async def fetch_chain_subaccount_orders(
+        self,
+        subaccount_id: str,
+        market_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_subaccount_orders(
+            subaccount_id=subaccount_id,
+            market_id=market_id,
+        )
+
+    async def fetch_chain_trader_spot_transient_orders(
+        self,
+        market_id: str,
+        subaccount_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_trader_spot_transient_orders(
+            market_id=market_id,
+            subaccount_id=subaccount_id,
+        )
+
+    async def fetch_spot_mid_price_and_tob(
+        self,
+        market_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_spot_mid_price_and_tob(
+            market_id=market_id,
+        )
+
+    async def fetch_derivative_mid_price_and_tob(
+        self,
+        market_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_derivative_mid_price_and_tob(
+            market_id=market_id,
+        )
+
+    async def fetch_chain_derivative_orderbook(
+        self,
+        market_id: str,
+        limit_cumulative_notional: Optional[str] = None,
+        pagination: Optional[PaginationOption] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_derivative_orderbook(
+            market_id=market_id,
+            limit_cumulative_notional=limit_cumulative_notional,
+            pagination=pagination,
+        )
+
+    async def fetch_chain_trader_derivative_orders(
+        self,
+        market_id: str,
+        subaccount_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_trader_derivative_orders(
+            market_id=market_id,
+            subaccount_id=subaccount_id,
+        )
+
+    async def fetch_chain_account_address_derivative_orders(
+        self,
+        market_id: str,
+        account_address: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_account_address_derivative_orders(
+            market_id=market_id,
+            account_address=account_address,
+        )
+
+    async def fetch_chain_derivative_orders_by_hashes(
+        self,
+        market_id: str,
+        subaccount_id: str,
+        order_hashes: List[str],
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_derivative_orders_by_hashes(
+            market_id=market_id,
+            subaccount_id=subaccount_id,
+            order_hashes=order_hashes,
+        )
+
+    async def fetch_chain_trader_derivative_transient_orders(
+        self,
+        market_id: str,
+        subaccount_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_trader_derivative_transient_orders(
+            market_id=market_id,
+            subaccount_id=subaccount_id,
+        )
+
+    async def fetch_chain_derivative_markets(
+        self,
+        status: Optional[str] = None,
+        market_ids: Optional[List[str]] = None,
+        with_mid_price_and_tob: Optional[bool] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_derivative_markets(
+            status=status,
+            market_ids=market_ids,
+            with_mid_price_and_tob=with_mid_price_and_tob,
+        )
+
+    async def fetch_chain_derivative_market(
+        self,
+        market_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_derivative_market(
+            market_id=market_id,
+        )
+
+    async def fetch_derivative_market_address(self, market_id: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_derivative_market_address(market_id=market_id)
+
+    async def fetch_subaccount_trade_nonce(self, subaccount_id: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_subaccount_trade_nonce(subaccount_id=subaccount_id)
+
+    async def fetch_chain_positions(self) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_positions()
+
+    async def fetch_chain_subaccount_positions(self, subaccount_id: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_subaccount_positions(subaccount_id=subaccount_id)
+
+    async def fetch_chain_subaccount_position_in_market(self, subaccount_id: str, market_id: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_subaccount_position_in_market(
+            subaccount_id=subaccount_id,
+            market_id=market_id,
+        )
+
+    async def fetch_chain_subaccount_effective_position_in_market(
+        self, subaccount_id: str, market_id: str
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_subaccount_effective_position_in_market(
+            subaccount_id=subaccount_id,
+            market_id=market_id,
+        )
+
+    async def fetch_chain_perpetual_market_info(self, market_id: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_perpetual_market_info(market_id=market_id)
+
+    async def fetch_chain_expiry_futures_market_info(self, market_id: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_expiry_futures_market_info(market_id=market_id)
+
+    async def fetch_chain_perpetual_market_funding(self, market_id: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_perpetual_market_funding(market_id=market_id)
+
+    async def fetch_subaccount_order_metadata(self, subaccount_id: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_subaccount_order_metadata(subaccount_id=subaccount_id)
+
+    async def fetch_trade_reward_points(
+        self,
+        accounts: Optional[List[str]] = None,
+        pending_pool_timestamp: Optional[int] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_trade_reward_points(
+            accounts=accounts,
+            pending_pool_timestamp=pending_pool_timestamp,
+        )
+
+    async def fetch_pending_trade_reward_points(
+        self,
+        accounts: Optional[List[str]] = None,
+        pending_pool_timestamp: Optional[int] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_pending_trade_reward_points(
+            accounts=accounts,
+            pending_pool_timestamp=pending_pool_timestamp,
+        )
+
+    async def fetch_trade_reward_campaign(self) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_trade_reward_campaign()
+
+    async def fetch_fee_discount_account_info(self, account: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_fee_discount_account_info(account=account)
+
+    async def fetch_fee_discount_schedule(self) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_fee_discount_schedule()
+
+    async def fetch_balance_mismatches(self, dust_factor: int) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_balance_mismatches(dust_factor=dust_factor)
+
+    async def fetch_balance_with_balance_holds(self) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_balance_with_balance_holds()
+
+    async def fetch_fee_discount_tier_statistics(self) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_fee_discount_tier_statistics()
+
+    async def fetch_mito_vault_infos(self) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_mito_vault_infos()
+
+    async def fetch_market_id_from_vault(self, vault_address: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_market_id_from_vault(vault_address=vault_address)
+
+    async def fetch_historical_trade_records(self, market_id: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_historical_trade_records(market_id=market_id)
+
+    async def fetch_is_opted_out_of_rewards(self, account: str) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_is_opted_out_of_rewards(account=account)
+
+    async def fetch_opted_out_of_rewards_accounts(self) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_opted_out_of_rewards_accounts()
+
+    async def fetch_market_volatility(
+        self,
+        market_id: str,
+        trade_grouping_sec: Optional[int] = None,
+        max_age: Optional[int] = None,
+        include_raw_history: Optional[bool] = None,
+        include_metadata: Optional[bool] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_market_volatility(
+            market_id=market_id,
+            trade_grouping_sec=trade_grouping_sec,
+            max_age=max_age,
+            include_raw_history=include_raw_history,
+            include_metadata=include_metadata,
+        )
+
+    async def fetch_chain_binary_options_markets(self, status: Optional[str] = None) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_binary_options_markets(status=status)
+
+    async def fetch_trader_derivative_conditional_orders(
+        self,
+        subaccount_id: Optional[str] = None,
+        market_id: Optional[str] = None,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_trader_derivative_conditional_orders(
+            subaccount_id=subaccount_id,
+            market_id=market_id,
+        )
+
+    async def fetch_market_atomic_execution_fee_multiplier(
+        self,
+        market_id: str,
+    ) -> Dict[str, Any]:
+        return await self.chain_exchange_api.fetch_market_atomic_execution_fee_multiplier(
+            market_id=market_id,
+        )
 
     # Injective Exchange client methods
 
@@ -2659,22 +3121,13 @@ class AsyncClient:
         )
 
         for market_info in valid_markets:
-            ticker = market_info["ticker"]
-            if "/" in ticker:
-                base_token_symbol, quote_token_symbol = ticker.split(constant.TICKER_TOKENS_SEPARATOR)
-            else:
-                base_token_symbol = market_info["baseTokenMeta"]["symbol"]
-                quote_token_symbol = market_info["quoteTokenMeta"]["symbol"]
-
             base_token = self._token_representation(
-                symbol=base_token_symbol,
                 token_meta=market_info["baseTokenMeta"],
                 denom=market_info["baseDenom"],
                 tokens_by_denom=tokens_by_denom,
                 tokens_by_symbol=tokens_by_symbol,
             )
             quote_token = self._token_representation(
-                symbol=quote_token_symbol,
                 token_meta=market_info["quoteTokenMeta"],
                 denom=market_info["quoteDenom"],
                 tokens_by_denom=tokens_by_denom,
@@ -2704,10 +3157,7 @@ class AsyncClient:
         )
 
         for market_info in valid_markets:
-            quote_token_symbol = market_info["quoteTokenMeta"]["symbol"]
-
             quote_token = self._token_representation(
-                symbol=quote_token_symbol,
                 token_meta=market_info["quoteTokenMeta"],
                 denom=market_info["quoteDenom"],
                 tokens_by_denom=tokens_by_denom,
@@ -2769,7 +3219,6 @@ class AsyncClient:
 
     def _token_representation(
         self,
-        symbol: str,
         token_meta: Dict[str, Any],
         denom: str,
         tokens_by_denom: Dict[str, Token],
@@ -2777,14 +3226,14 @@ class AsyncClient:
     ) -> Token:
         if denom not in tokens_by_denom:
             unique_symbol = denom
-            for symbol_candidate in [symbol, token_meta["symbol"], token_meta["name"]]:
+            for symbol_candidate in [token_meta["symbol"], token_meta["name"]]:
                 if symbol_candidate not in tokens_by_symbol:
                     unique_symbol = symbol_candidate
                     break
 
             token = Token(
                 name=token_meta["name"],
-                symbol=symbol,
+                symbol=token_meta["symbol"],
                 denom=denom,
                 address=token_meta["address"],
                 decimals=token_meta["decimals"],